+++ src/auth/stscredentials.ts
@@ -122,7 +122,7 @@ export interface StsSuccessfulResponse {
   token_type: string;
   expires_in: number;
   refresh_token?: string;
-  scope: string;
+  scope?: string;
   res?: GaxiosResponse | null;
 }